+++ packages/gitmoji-regex/.changelogrc.js
@@ -1,0 +1,6 @@
+const base = require('../../.changelogrc');
+
+module.exports = {
+  ...base,
+  displayScopes: ['gitmoji-regex'],
+};